Put get_ibp(mode) and get_affine near get_mode(ibp, affine)
