+++ src/decomon/backward_layers/deel_lip.py
@@ -4,8 +4,7 @@ from typing import Any, Dict, Optional, Union
 from tensorflow.keras.layers import Layer
 
 from decomon.backward_layers.core import BackwardLayer
-from decomon.backward_layers.utils import get_affine, get_ibp
-from decomon.core import ForwardMode, PerturbationDomain
+from decomon.core import ForwardMode, PerturbationDomain, get_affine, get_ibp
 from decomon.layers.convert import to_decomon
 from decomon.layers.decomon_layers import DecomonLayer
