+++ tests/test_dense_layer.py
@@ -7,8 +7,7 @@ import tensorflow.keras.backend as K
 from numpy.testing import assert_almost_equal
 from tensorflow.keras.layers import Dense
 
-from decomon.backward_layers.utils import get_affine, get_ibp
-from decomon.core import ForwardMode
+from decomon.core import ForwardMode, get_affine, get_ibp
 from decomon.layers.convert import to_decomon
 from decomon.layers.decomon_layers import DecomonDense
 from decomon.models.utils import split_activation
