+++ tests/test_merge.py
@@ -15,8 +15,7 @@ from tensorflow.keras.layers import (
 )
 from tensorflow.keras.models import Model
 
-from decomon.backward_layers.utils import get_affine, get_ibp
-from decomon.core import ForwardMode
+from decomon.core import ForwardMode, get_affine, get_ibp
 from decomon.layers.convert import to_decomon
 from decomon.layers.decomon_merge_layers import (
     DecomonAdd,
